More forward declarations around S3/Azure/IObjectStorage/IDisk
The most important is around IObjectStorage.h/IDisk.h
