+++ src/Backups/BackupEntriesCollector.cpp
@@ -19,6 +19,7 @@
 #include <base/scope_guard.h>
 #include <base/sleep.h>
 #include <Common/escapeForFileName.h>
+#include <Common/threadPoolCallbackRunner.h>
 #include <Core/Settings.h>
 
 #include <boost/range/adaptor/map.hpp>
