+++ src/Backups/BackupFileInfo.cpp
@@ -1,5 +1,4 @@
 #include <Backups/BackupFileInfo.h>
-
 #include <Backups/IBackup.h>
 #include <Backups/IBackupEntry.h>
 #include <Common/CurrentThread.h>
@@ -7,6 +6,7 @@
 #include <Common/scope_guard_safe.h>
 #include <Common/setThreadName.h>
 #include <Common/ThreadPool.h>
+#include <Common/threadPoolCallbackRunner.h>
 #include <Interpreters/ProcessList.h>
 
 #include <base/hex.h>
