+++ src/Coordination/KeeperCommon.cpp
@@ -2,6 +2,7 @@
 
 #include <string>
 #include <filesystem>
+#include <thread>
 
 #include <Common/logger_useful.h>
 #include <Disks/IDisk.h>
