+++ src/Interpreters/ServerAsynchronousMetrics.cpp
@@ -17,6 +17,11 @@
 #include <IO/UncompressedCache.h>
 #include <IO/MMappedFileCache.h>
 
+#include "config.h"
+#if USE_AWS_S3
+#include <IO/S3/Client.h>
+#endif
+
 #include <Storages/MergeTree/MergeTreeData.h>
 #include <Storages/StorageMergeTree.h>
 #include <Storages/StorageReplicatedMergeTree.h>
