+++ src/Interpreters/tests/gtest_filecache.cpp
@@ -1,3 +1,4 @@
+#include <IO/copyData.h>
 #include <gtest/gtest.h>
 
 #include <filesystem>
@@ -5,9 +6,7 @@
 
 
 #include <algorithm>
-#include <numeric>
 #include <thread>
-#include <chrono>
 
 #include <Core/ServerUUID.h>
 #include <Common/iota.h>
