+++ src/Storages/RabbitMQ/StorageRabbitMQ.cpp
@@ -33,6 +33,7 @@
 #include <Common/parseAddress.h>
 #include <Common/quoteString.h>
 #include <Common/setThreadName.h>
+#include <Common/RemoteHostFilter.h>
 
 #include <base/range.h>
